(python) jawbreaker algorithm.
this algorithm basically works by taking an input and attempting to
guess it with a certain amount of generations
that can occur at one time. This is the slower but more accurate and
"Industry" password breaker, which can be used
for other things and serves as a more realistic password breaker.